+++ src/emp_model/src/weather.cpp
@@ -32,7 +32,7 @@ bool Weather::fillFromJSON(json * j)
         data.push_back( HourlyData() );
 		data[i].month = d[i].at("month").get<int>();
 		data[i].day = d[i].at("day").get<int>();
-		data[i].hour = d[i].at("hour").get<double>();
+		data[i].hour = d[i].at("hour").get<float>();
 		data[i].diffuse_horizontal = d[i].at("diffuse_horizontal").get<int>();
 		data[i].direct_normal = d[i].at("direct_normal").get<int>();
 	}
